refactor(cmd): decompose serve_handlers.go into per-tool files [godfile-serve-handlers]
Mechanical decomposition of cmd/serve_handlers.go (1374 → 182 LOC) into 10 per-tool handler files + 8 behavioral tests for residual paths. Coverage-gate exit 0. Skeptic verdict APPROVE WITH COMMENT; all findings addressed (rosary-2b21a3 isolation, 9 stale-line-refs rewritten, sun_path constraint documented). Closes godfile-serve-handlers.
