+++ cmd/serve_handlers.go
@@ -1,37 +1,9 @@
 package cmd
 
 import (
-	"context"
-	"database/sql"
-	"encoding/json"
-	"fmt"
-	"io"
-	"path/filepath"
-	"sort"
-	"strings"
-
-	"github.com/agentic-research/mache/internal/graph"
-	"github.com/agentic-research/mache/internal/leyline"
 	"github.com/mark3labs/mcp-go/mcp"
 	"github.com/mark3labs/mcp-go/server"
 )
-
-// prioritizeAndAdvise sends a fire-and-forget hint to the ley-line daemon
-// to prioritize parsing the given files. Returns a user-facing message, or
-// empty string if not in daemon mode or the daemon is unreachable.
-func prioritizeAndAdvise(files []string) string {
-	if serveControl == "" || len(files) == 0 {
-		return ""
-	}
-	sockPath := strings.TrimSuffix(serveControl, ".ctrl") + ".sock"
-	sc, err := leyline.DialSocket(sockPath)
-	if err != nil {
-		return ""
-	}
-	defer func() { _ = sc.Close() }()
-	_ = sc.Prioritize(files) // fire-and-forget
-	return "File not yet parsed — prioritized for next daemon pass, retry shortly."
-}
 
 // registerMCPTools registers all tool definitions with session-aware handlers.
 // All tools are registered unconditionally — lazyGraph delegates to the inner
@@ -207,1168 +179,4 @@ func registerMCPTools(s *server.MCPServer, r *graphRegistry) {
 		),
 		r.wrapHandler(makeFindSmellsHandler),
 	)
-}
-
-type nodeEntry struct {
-	Name string `json:"name"`
-	Path string `json:"path"`
-	Type string `json:"type"`
-	Size int64  `json:"size,omitempty"`
-}
-
-// containerNames are schema-level dir names that group constructs but are
-// never themselves a callable token. Skipping the GetCallers/GetCallees
-// virtual-entry check at these paths avoids an O(N) SQL query and a
-// tree-sitter parse per directory listing — the dominant cost on large
-// monorepos (bead mache-og26).
-//
-// Adding a new schema container here is safe; the worst case is missing a
-// virtual `callers/` or `callees/` entry on a dir that wasn't going to
-// have meaningful refs anyway.
-var containerNames = map[string]bool{
-	"methods":    true,
-	"functions":  true,
-	"types":      true,
-	"interfaces": true,
-	"structs":    true,
-	"classes":    true,
-	"imports":    true,
-	"constants":  true,
-	"variables":  true,
-	"namespaces": true,
-	"enums":      true,
-	"protocols":  true,
-	"objects":    true,
-	"traits":     true,
-	"exports":    true,
-	"fields":     true,
-}
-
-func makeListDirHandler(g graph.Graph) server.ToolHandlerFunc {
-	return func(_ context.Context, request mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-		path := request.GetString("path", "")
-		excludeTests := request.GetBool("exclude_tests", false)
-
-		// Single-RLock snapshot of all children. Replaces the prior
-		// pattern of ListChildren + N×GetNode (one lookup per child).
-		// On a large dir (300+ children) this is the difference between
-		// one map iteration and 300 lock acquisitions.
-		stats, err := g.ListChildStats(path)
-		if err != nil {
-			return mcp.NewToolResultError(fmt.Sprintf("list %q: %v", path, err)), nil
-		}
-
-		entries := make([]nodeEntry, 0, len(stats))
-		for _, s := range stats {
-			name := filepath.Base(s.ID)
-			if excludeTests && (strings.HasPrefix(name, "Test") || strings.HasPrefix(name, "Benchmark")) {
-				continue
-			}
-			typ := "file"
-			if s.IsDir {
-				typ = "dir"
-			}
-			entries = append(entries, nodeEntry{
-				Name: name,
-				Path: s.ID,
-				Type: typ,
-				Size: s.ContentSize,
-			})
-		}
-
-		// Surface virtual entries on construct directories
-		// (skip if already materialized in the db)
-		if path != "" {
-			seen := make(map[string]bool, len(entries))
-			for _, e := range entries {
-				seen[e.Name] = true
-			}
-
-			// Location: source file coordinates for orientation
-			if !seen["location"] {
-				if node, err := g.GetNode(path); err == nil && node.Properties != nil {
-					if loc, ok := node.Properties["location"]; ok && len(loc) > 0 {
-						entries = append(entries, nodeEntry{
-							Name: "location",
-							Path: path + "/location",
-							Type: "virtual",
-							Size: int64(len(loc)),
-						})
-					}
-				}
-			}
-
-			// Skip GetCallers / GetCallees probes at schema container
-			// levels (methods/, functions/, types/, ...). Their basename
-			// is never a callable token, so the SQL/parse cost is wasted
-			// every readdir.
-			token := filepath.Base(path)
-			isContainer := containerNames[token]
-
-			if !isContainer && !seen["callers"] {
-				if callers, err := g.GetCallers(token); err == nil && len(callers) > 0 {
-					entries = append(entries, nodeEntry{
-						Name: "callers",
-						Path: path + "/callers",
-						Type: "virtual",
-					})
-				}
-			}
-			if !isContainer && !seen["callees"] {
-				if callees, err := g.GetCallees(path); err == nil && len(callees) > 0 {
-					entries = append(entries, nodeEntry{
-						Name: "callees",
-						Path: path + "/callees",
-						Type: "virtual",
-					})
-				}
-			}
-		}
-
-		data, _ := json.MarshalIndent(entries, "", "  ")
-		return mcp.NewToolResultText(string(data)), nil
-	}
-}
-
-const maxReadFileSize = 32 * 1024 * 1024 // 32 MB per file
-
-type fileReadResult struct {
-	Content string              `json:"content"`
-	Origin  *graph.SourceOrigin `json:"origin,omitempty"`
-}
-
-func readOneFileWithOrigin(g graph.Graph, path string) (*fileReadResult, error) {
-	// Handle virtual location file
-	if filepath.Base(path) == graph.LocationFile {
-		parentDir := filepath.Dir(path)
-		if parent, err := g.GetNode(parentDir); err == nil && parent.Properties != nil {
-			if loc, ok := parent.Properties["location"]; ok && len(loc) > 0 {
-				return &fileReadResult{Content: string(loc)}, nil
-			}
-		}
-		return nil, fmt.Errorf("not found: %s", path)
-	}
-
-	node, err := g.GetNode(path)
-	if err != nil {
-		return nil, fmt.Errorf("not found: %s", path)
-	}
-	if node.Mode.IsDir() {
-		return nil, fmt.Errorf("%s is a directory — use list_directory", path)
-	}
-	size := node.ContentSize()
-	if size == 0 {
-		return &fileReadResult{Origin: node.Origin}, nil
-	}
-	if size > maxReadFileSize {
-		return nil, fmt.Errorf("%s too large (%d bytes, max %d)", path, size, maxReadFileSize)
-	}
-	buf := make([]byte, size)
-	n, err := g.ReadContent(path, buf, 0)
-	if err != nil {
-		return nil, fmt.Errorf("read %s: %v", path, err)
-	}
-	return &fileReadResult{Content: string(buf[:n]), Origin: node.Origin}, nil
-}
-
-func makeReadFileHandler(g graph.Graph) server.ToolHandlerFunc {
-	return func(ctx context.Context, request mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-		path := request.GetString("path", "")
-		pathsRaw := request.GetString("paths", "")
-
-		// Batch mode
-		if pathsRaw != "" {
-			var paths []string
-			if err := json.Unmarshal([]byte(pathsRaw), &paths); err != nil {
-				return mcp.NewToolResultError(fmt.Sprintf("invalid paths array: %v", err)), nil
-			}
-			const maxBatchPaths = 100
-			if len(paths) > maxBatchPaths {
-				return mcp.NewToolResultError(fmt.Sprintf("batch read limited to %d paths, got %d", maxBatchPaths, len(paths))), nil
-			}
-
-			type fileResult struct {
-				Path    string              `json:"path"`
-				Content string              `json:"content,omitempty"`
-				Origin  *graph.SourceOrigin `json:"origin,omitempty"`
-				Error   string              `json:"error,omitempty"`
-			}
-			const maxBatchBytes int64 = maxReadFileSize // total content cap for batch
-			results := make([]fileResult, 0, len(paths))
-			var totalBytes int64
-			for i, p := range paths {
-				// Pre-check size to avoid allocating before rejecting.
-				if node, err := g.GetNode(p); err == nil && !node.Mode.IsDir() {
-					totalBytes += node.ContentSize()
-					if totalBytes > maxBatchBytes {
-						results = append(results, fileResult{Path: p, Error: fmt.Sprintf("batch too large (exceeds %d bytes total)", maxBatchBytes)})
-						for _, remaining := range paths[i+1:] {
-							results = append(results, fileResult{Path: remaining, Error: "skipped: batch size limit reached"})
-						}
-						break
-					}
-				}
-				r, err := readOneFileWithOrigin(g, p)
-				if err != nil {
-					results = append(results, fileResult{Path: p, Error: err.Error()})
-					continue
-				}
-				results = append(results, fileResult{Path: p, Content: r.Content, Origin: r.Origin})
-			}
-			data, _ := json.MarshalIndent(results, "", "  ")
-			return mcp.NewToolResultText(string(data)), nil
-		}
-
-		// Single mode
-		if path == "" {
-			return mcp.NewToolResultError("path or paths is required"), nil
-		}
-		r, err := readOneFileWithOrigin(g, path)
-		if err != nil {
-			// In daemon mode, the file may not be parsed yet — ask daemon to prioritize it.
-			if msg := prioritizeAndAdvise([]string{path}); msg != "" {
-				return mcp.NewToolResultText(msg), nil
-			}
-			return mcp.NewToolResultError(err.Error()), nil
-		}
-		// If there's an origin, return it as structured JSON so the consumer
-		// knows exactly where to edit in the real filesystem.
-		if r.Origin != nil {
-			data, _ := json.MarshalIndent(r, "", "  ")
-			return mcp.NewToolResultText(string(data)), nil
-		}
-		return mcp.NewToolResultText(r.Content), nil
-	}
-}
-
-func makeFindCallersHandler(g graph.Graph) server.ToolHandlerFunc {
-	return func(ctx context.Context, request mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-		token := request.GetString("token", "")
-		if token == "" {
-			return mcp.NewToolResultError("token is required"), nil
-		}
-
-		callers, err := g.GetCallers(token)
-		if err != nil {
-			return mcp.NewToolResultError(fmt.Sprintf("get callers: %v", err)), nil
-		}
-
-		paths := make([]string, 0, len(callers))
-		for _, c := range callers {
-			paths = append(paths, c.ID)
-		}
-
-		// Cross-repo annotation: when running on a CompositeGraph,
-		// surface each caller's mount of origin. annotateMounts
-		// returns nil if there's nothing useful to annotate; the
-		// handler then falls through to the legacy []string shape.
-		if scoped := annotateMounts(g, paths); scoped != nil {
-			type callersResult struct {
-				Callers []scopedItem `json:"callers"`
-			}
-			data, _ := json.MarshalIndent(callersResult{Callers: scoped}, "", "  ")
-			return mcp.NewToolResultText(string(data)), nil
-		}
-
-		// Supplement with LSP references if available
-		if qg, ok := g.(refsQuerier); ok {
-			lspRefs, lspErr := queryLSPRefs(qg, token)
-			if lspErr == nil && len(lspRefs) > 0 {
-				type callersResult struct {
-					Callers []string         `json:"callers"`
-					LSPRefs []lspRefLocation `json:"lsp_refs"`
-				}
-				data, _ := json.MarshalIndent(callersResult{
-					Callers: paths,
-					LSPRefs: lspRefs,
-				}, "", "  ")
-				return mcp.NewToolResultText(string(data)), nil
-			}
-		}
-
-		// No LSP data — return original format for backward compatibility
-		if len(paths) == 0 {
-			if serveControl != "" {
-				return mcp.NewToolResultText("[] — daemon may still be parsing, retry shortly"), nil
-			}
-			return mcp.NewToolResultText("[]"), nil
-		}
-		data, _ := json.MarshalIndent(paths, "", "  ")
-		return mcp.NewToolResultText(string(data)), nil
-	}
-}
-
-// genericGoNames is the set of method/function names that commonly cause
-// false positives in bare-token callee resolution. When a callee's base name
-// matches one of these, the result is flagged as potentially noisy.
-var genericGoNames = map[string]bool{
-	"String": true, "Error": true, "New": true, "Parse": true,
-	"Close": true, "Read": true, "Write": true, "Open": true,
-	"Run": true, "Start": true, "Stop": true, "Reset": true,
-	"Marshal": true, "Unmarshal": true, "Encode": true, "Decode": true,
-	"Format": true, "Scan": true, "Next": true, "Done": true,
-}
-
-func makeFindCalleesHandler(g graph.Graph) server.ToolHandlerFunc {
-	return func(ctx context.Context, request mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-		path := request.GetString("path", "")
-		if path == "" {
-			return mcp.NewToolResultError("path is required"), nil
-		}
-
-		callees, err := g.GetCallees(path)
-		if err != nil {
-			return mcp.NewToolResultError(fmt.Sprintf("get callees: %v", err)), nil
-		}
-		if len(callees) == 0 {
-			// Provide a helpful hint about why callees might be empty
-			node, nodeErr := g.GetNode(path)
-			if nodeErr != nil {
-				return mcp.NewToolResultText(`{"callees":[],"hint":"construct not found — check the path"}`), nil
-			}
-			if !node.Mode.IsDir() {
-				return mcp.NewToolResultText(`{"callees":[],"hint":"path is a file, not a construct directory — use the parent directory path"}`), nil
-			}
-			hint := "no resolved callees"
-			if node.Properties != nil {
-				if _, hasLang := node.Properties["lang"]; hasLang {
-					hint = "no resolved callees — the construct may call unexported methods or use dynamic dispatch that the static extractor cannot resolve. Try find_callers with the method name instead."
-				}
-			}
-			type emptyResult struct {
-				Callees []string `json:"callees"`
-				Hint    string   `json:"hint"`
-			}
-			data, _ := json.MarshalIndent(emptyResult{Callees: []string{}, Hint: hint}, "", "  ")
-			return mcp.NewToolResultText(string(data)), nil
-		}
-
-		calleePaths := make([]string, 0, len(callees))
-		var warnings []string
-		seenGeneric := make(map[string]bool)
-		for _, c := range callees {
-			calleePaths = append(calleePaths, c.ID)
-			// Check if this callee's base name is a common generic identifier.
-			// Generic names resolved via bare-token fallback may include false
-			// positives from unrelated packages.
-			name := filepath.Base(c.ID)
-			if genericGoNames[name] && !seenGeneric[name] {
-				seenGeneric[name] = true
-				warnings = append(warnings,
-					fmt.Sprintf("'%s' is a common name — results may include false positives from unrelated packages. Use find_callers on the specific implementation path to verify.", name),
-				)
-			}
-		}
-
-		// Cross-repo annotation — same pattern as find_callers.
-		if scoped := annotateMounts(g, calleePaths); scoped != nil {
-			type calleesResult struct {
-				Callees  []scopedItem `json:"callees"`
-				Warnings []string     `json:"warnings,omitempty"`
-			}
-			out := calleesResult{Callees: scoped, Warnings: warnings}
-			data, _ := json.MarshalIndent(out, "", "  ")
-			return mcp.NewToolResultText(string(data)), nil
-		}
-
-		type calleesResult struct {
-			Callees  []string `json:"callees"`
-			Warnings []string `json:"warnings,omitempty"`
-		}
-		out := calleesResult{Callees: calleePaths, Warnings: warnings}
-		data, _ := json.MarshalIndent(out, "", "  ")
-		return mcp.NewToolResultText(string(data)), nil
-	}
-}
-
-func makeSearchHandler(g graph.Graph) server.ToolHandlerFunc {
-	return func(ctx context.Context, request mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-		pattern := request.GetString("pattern", "")
-		if pattern == "" {
-			return mcp.NewToolResultError("pattern is required"), nil
-		}
-
-		typeFilter := request.GetString("type", "")
-		role := request.GetString("role", "")
-		limit := request.GetInt("limit", 100)
-
-		type searchResult struct {
-			Token string `json:"token"`
-			Path  string `json:"path"`
-			Role  string `json:"role,omitempty"`
-		}
-
-		// Definition search.
-		//
-		// Two backends: SQL-pushdown via defsSearcher (preferred — covers
-		// pre-built .db files whose in-memory defs map is empty), and
-		// in-memory map iteration via defsMapProvider (legacy MemoryStore
-		// and live-ingest paths).
-		//
-		// Wrappers like lazyGraph always satisfy the defsSearcher
-		// interface (they have the method, even if the inner backend
-		// doesn't). If the wrapper returns nil from SearchDefs because
-		// the inner didn't implement it, fall through to the DefsMap
-		// path — `nil` from SearchDefs is distinguishable from `empty
-		// map = no matches` and signals "I don't speak this protocol."
-		//
-		// Bead mache-9cba08 caught the original SQL pushdown gap; the
-		// nil-passthrough fallback below was caught by the PR #373
-		// post-fix audit (lazyGraph→MemoryStore returned nil and the
-		// fallback was unreachable due to dispatch ordering).
-		if role == "definition" {
-			var matches map[string][]string
-			if ds, ok := g.(defsSearcher); ok {
-				matches = ds.SearchDefs(pattern, limit)
-			}
-			if matches == nil {
-				if dp, ok := g.(defsMapProvider); ok {
-					defs := dp.DefsMap()
-					matches = make(map[string][]string, len(defs))
-					for token, ids := range defs {
-						if sqlLikeMatch(pattern, token) {
-							matches[token] = ids
-						}
-					}
-				} else {
-					return mcp.NewToolResultError("backend does not support definition search"), nil
-				}
-			}
-
-			var results []searchResult
-			seenPaths := make(map[string]bool)
-			for token, ids := range matches {
-				for _, id := range ids {
-					if typeFilter != "" && !strings.Contains(id, "/"+typeFilter+"/") {
-						continue
-					}
-					// Dedup: both "Foo.Bar" and "pkg.Foo.Bar" map to the same path
-					if seenPaths[id] {
-						continue
-					}
-					seenPaths[id] = true
-					results = append(results, searchResult{Token: token, Path: id, Role: "definition"})
-					if len(results) >= limit {
-						break
-					}
-				}
-				if len(results) >= limit {
-					break
-				}
-			}
-			if results == nil {
-				results = []searchResult{}
-			}
-			data, _ := json.MarshalIndent(results, "", "  ")
-			return mcp.NewToolResultText(string(data)), nil
-		}
-
-		// Reference search (default): query mache_refs (vtab) or node_refs (real table).
-		// Leyline-parsed .dbs have node_refs (token, node_id); legacy sidecar has mache_refs (token, path).
-		//
-		// `path NOT LIKE '_file_level:%'` filters the engine's
-		// file-level sentinel rows (from PR #270) — synthetic
-		// caller_ids the engine uses to mark file-level fn-value
-		// refs (mache-02r9) as alive. They're not real source
-		// locations and would surface as fake "/path/file.go"
-		// results. NodesTableReader.GetCallers applies the same
-		// filter; this puts search/role=reference in agreement.
-		qg, ok := g.(refsQuerier)
-		if !ok {
-			return mcp.NewToolResultError("reference search requires a SQLite-backed graph; use role=definition for in-memory search"), nil
-		}
-		var rows *sql.Rows
-		var err error
-		if typeFilter != "" {
-			rows, err = qg.QueryRefs(
-				"SELECT token, path FROM mache_refs WHERE token LIKE ? AND path LIKE ? AND path NOT LIKE '_file_level:%' LIMIT ?",
-				pattern, "%/"+typeFilter+"/%", limit,
-			)
-		} else {
-			rows, err = qg.QueryRefs(
-				"SELECT token, path FROM mache_refs WHERE token LIKE ? AND path NOT LIKE '_file_level:%' LIMIT ?",
-				pattern, limit,
-			)
-		}
-		// Fallback: node_refs table (leyline-parsed .db) when mache_refs vtab doesn't exist.
-		if err != nil && strings.Contains(err.Error(), "no such table") {
-			if typeFilter != "" {
-				rows, err = qg.QueryRefs(
-					"SELECT token, node_id AS path FROM node_refs WHERE token LIKE ? AND node_id LIKE ? AND node_id NOT LIKE '_file_level:%' LIMIT ?",
-					pattern, "%/"+typeFilter+"/%", limit,
-				)
-			} else {
-				rows, err = qg.QueryRefs(
-					"SELECT token, node_id AS path FROM node_refs WHERE token LIKE ? AND node_id NOT LIKE '_file_level:%' LIMIT ?",
-					pattern, limit,
-				)
-			}
-		}
-		if err != nil {
-			return mcp.NewToolResultError(fmt.Sprintf("query: %v", err)), nil
-		}
-		defer func() { _ = rows.Close() }()
-
-		var results []searchResult
-		for rows.Next() {
-			var r searchResult
-			if err := rows.Scan(&r.Token, &r.Path); err != nil {
-				continue
-			}
-			results = append(results, r)
-		}
-		if results == nil {
-			results = []searchResult{}
-		}
-
-		data, _ := json.MarshalIndent(results, "", "  ")
-		return mcp.NewToolResultText(string(data)), nil
-	}
-}
-
-func makeSemanticSearchHandler(g graph.Graph) server.ToolHandlerFunc {
-	return func(ctx context.Context, request mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-		query := request.GetString("query", "")
-		if query == "" {
-			return mcp.NewToolResultError("query is required"), nil
-		}
-		k := request.GetInt("k", 10)
-
-		sockPath, err := leyline.DiscoverOrStart()
-		if err != nil {
-			return mcp.NewToolResultError(
-				"semantic search not available — requires ley-line daemon with embeddings.\n" +
-					"This is an optional feature. Use 'search' for pattern-based code search instead.",
-			), nil
-		}
-
-		sock, err := leyline.DialSocket(sockPath)
-		if err != nil {
-			return mcp.NewToolResultError(
-				"semantic search not available — ley-line daemon not responding.\n" +
-					"This is an optional feature. Use 'search' for pattern-based code search instead.",
-			), nil
-		}
-		defer func() { _ = sock.Close() }()
-
-		sc := leyline.NewSemanticClient(sock)
-		results, err := sc.Search(query, k)
-		if err != nil {
-			return mcp.NewToolResultError(
-				"semantic search not available — ley-line daemon does not support embeddings.\n" +
-					"This is an optional feature. Use 'search' for pattern-based code search instead.",
-			), nil
-		}
-
-		if len(results) == 0 {
-			return mcp.NewToolResultText("[]"), nil
-		}
-
-		type enrichedResult struct {
-			Path     string  `json:"path"`
-			Distance float64 `json:"distance"`
-			Type     string  `json:"type,omitempty"`
-			Snippet  string  `json:"snippet,omitempty"`
-		}
-
-		enriched := make([]enrichedResult, 0, len(results))
-		for _, r := range results {
-			er := enrichedResult{
-				Path:     r.ID,
-				Distance: r.Distance,
-			}
-
-			// Enrich with graph metadata
-			node, nodeErr := g.GetNode(r.ID)
-			if nodeErr == nil && node != nil {
-				if node.Mode.IsDir() {
-					er.Type = "directory"
-				} else {
-					er.Type = "file"
-					// Read a content snippet (first 200 bytes)
-					buf := make([]byte, 200)
-					n, _ := g.ReadContent(r.ID, buf, 0)
-					if n > 0 {
-						snippet := string(buf[:n])
-						if n == 200 {
-							snippet += "..."
-						}
-						er.Snippet = snippet
-					}
-				}
-			}
-
-			enriched = append(enriched, er)
-		}
-
-		data, _ := json.MarshalIndent(enriched, "", "  ")
-		return mcp.NewToolResultText(string(data)), nil
-	}
-}
-
-// sqlLikeMatch performs a simple SQL LIKE pattern match (% = wildcard).
-func sqlLikeMatch(pattern, value string) bool {
-	pattern = strings.ToLower(pattern)
-	value = strings.ToLower(value)
-
-	// Fast paths for common patterns
-	if pattern == "%" {
-		return true
-	}
-	if !strings.Contains(pattern, "%") {
-		return pattern == value
-	}
-
-	parts := strings.Split(pattern, "%")
-	pos := 0
-	for i, part := range parts {
-		if part == "" {
-			continue
-		}
-		idx := strings.Index(value[pos:], part)
-		if idx < 0 {
-			return false
-		}
-		if i == 0 && idx != 0 {
-			// First part must match start if pattern doesn't begin with %
-			return false
-		}
-		pos += idx + len(part)
-	}
-	// If pattern doesn't end with %, value must end at pos
-	if !strings.HasSuffix(pattern, "%") && pos != len(value) {
-		return false
-	}
-	return true
-}
-
-func makeGetCommunitiesHandler(g graph.Graph) server.ToolHandlerFunc {
-	return makeGetCommunitiesHandlerWithDone(g, nil)
-}
-
-// makeGetCommunitiesHandlerWithDone is the test-friendly variant of
-// makeGetCommunitiesHandler. When pushDone is non-nil, the channel is
-// closed once the fire-and-forget PushTopology goroutine returns —
-// letting race-prone tests wait deterministically instead of sleeping
-// or polling. Production callers use the nil-channel path via the
-// public makeGetCommunitiesHandler wrapper.
-func makeGetCommunitiesHandlerWithDone(g graph.Graph, pushDone chan<- struct{}) server.ToolHandlerFunc {
-	return func(ctx context.Context, request mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-		minSize := request.GetInt("min_size", 2)
-		summary := request.GetBool("summary", false)
-
-		rp, ok := g.(refsMapProvider)
-		if !ok {
-			return mcp.NewToolResultError("community detection requires a graph with cross-reference data (SQLite backend or MemoryStore with refs)"), nil
-		}
-		refs := rp.RefsMap()
-		if len(refs) == 0 {
-			type emptyResult struct {
-				Communities []any   `json:"communities"`
-				NumNodes    int     `json:"num_nodes"`
-				NumEdges    int     `json:"num_edges"`
-				Modularity  float64 `json:"modularity"`
-				Message     string  `json:"message"`
-			}
-			data, _ := json.Marshal(emptyResult{
-				Communities: []any{},
-				Message: "No cross-references indexed. Community detection requires constructs that share symbols. " +
-					"Ensure the source was ingested with a schema that captures references (sitter_walker or json_walker with refs). " +
-					"Use get_overview to check ref_tokens count.",
-			})
-			return mcp.NewToolResultText(string(data)), nil
-		}
-
-		result := graph.DetectCommunities(refs, minSize)
-
-		// Push topology to ley-line sheaf cache (fire-and-forget).
-		// Errors are best-effort: the daemon may be missing
-		// (DiscoverOrStart fails), unreachable (DialSocket fails),
-		// or running an older protocol that doesn't support
-		// sheaf_set_topology. None of those are actionable for the
-		// user mid-call; log only once per process to avoid the
-		// every-call spam during benchmarks and non-LLO setups.
-		//
-		// Snapshot the inputs PushTopology reads (Communities + each
-		// community's Members) so the goroutine sees a stable view —
-		// the full-output branch below mutates result.Communities in
-		// place (sort + truncate slice + truncate per-community Members)
-		// and would otherwise race buildRegions inside PushTopology.
-		// Membership is read-only after DetectCommunities returns, so
-		// reusing it without cloning is safe.
-		snapComms := make([]graph.Community, len(result.Communities))
-		for i, c := range result.Communities {
-			snapComms[i] = graph.Community{
-				ID:      c.ID,
-				Members: append([]string(nil), c.Members...),
-			}
-		}
-		snap := &graph.CommunityResult{
-			Communities: snapComms,
-			Membership:  result.Membership,
-		}
-
-		// Resolve the SheafInvalidator (if this graph has one) up front
-		// so the goroutine below can install state atomically once it
-		// successfully reaches the daemon.
-		//
-		// Graph backends without a watcher (control-mode lazyGraph,
-		// composite mounts) don't implement sheafInvalidatorProvider —
-		// the type-assert degrades silently, which is the documented
-		// behavior for those modes (see buildServeGraph).
-		//
-		// DELIBERATELY DEFER state install until PushTopology succeeds.
-		// Installing SetCommunityResult synchronously here (the prior
-		// shape) paired NEW membership with the OLD sheaf/topology
-		// during the dial+push window — watcher events in that window
-		// looked up regions in the new membership but sent them to a
-		// daemon that still held the prior topology. Fixed by atomic
-		// SetState below; see PR #383 Copilot #6.
-		var sip sheafInvalidatorProvider
-		if sp, ok := g.(sheafInvalidatorProvider); ok {
-			sip = sp
-		}
-
-		go func() {
-			if pushDone != nil {
-				defer close(pushDone)
-			}
-			sockPath, err := leyline.DiscoverOrStart()
-			if err != nil {
-				return // no daemon — skip silently
-			}
-			sock, err := leyline.DialSocket(sockPath)
-			if err != nil {
-				return
-			}
-			// Track whether we handed off socket ownership to the
-			// invalidator. If we did, the SheafClient inside the
-			// invalidator owns the socket and we MUST NOT close it
-			// here — the watcher's later cascade calls reuse the
-			// connection. If we didn't hand off, close it normally
-			// to avoid leaking a socket per failed get_communities
-			// run. See PR #383 Copilot #1: the prior code deferred
-			// sock.Close() unconditionally, so the SheafClient
-			// stored on the invalidator was talking to a closed
-			// socket from the goroutine's return onward.
-			var handedOff bool
-			defer func() {
-				if !handedOff {
-					_ = sock.Close()
-				}
-			}()
-
-			sc := leyline.NewSheafClient(sock)
-			if pushErr := sc.PushTopology(snap, refs); pushErr != nil {
-				logSheafPushOnce(pushErr)
-				return
-			}
-			// PushTopology succeeded — atomically swap BOTH the new
-			// CommunityResult and the new SheafClient onto the
-			// invalidator. From here on, subsequent watcher fires
-			// engage the cross-region cascade. The prior backend
-			// (if any) is returned so we can close its socket and
-			// avoid leaking one UDS connection per get_communities
-			// call across the process lifetime.
-			if sip != nil {
-				if si := sip.SheafInvalidator(); si != nil {
-					prior := si.SetState(snap, sc)
-					handedOff = true
-					if closer, ok := prior.(io.Closer); ok {
-						_ = closer.Close()
-					}
-				}
-			}
-		}()
-
-		if summary {
-			type communitySummary struct {
-				ID         int      `json:"id"`
-				Size       int      `json:"size"`
-				TopMembers []string `json:"top_members"`
-			}
-			type summaryResult struct {
-				NumCommunities int                `json:"num_communities"`
-				NumNodes       int                `json:"num_nodes"`
-				NumEdges       int                `json:"num_edges"`
-				Modularity     float64            `json:"modularity"`
-				Communities    []communitySummary `json:"communities"`
-			}
-			sr := summaryResult{
-				NumCommunities: len(result.Communities),
-				NumNodes:       result.NumNodes,
-				NumEdges:       result.NumEdges,
-				Modularity:     result.Modularity,
-			}
-			for _, c := range result.Communities {
-				top := c.Members
-				if len(top) > 5 {
-					top = top[:5]
-				}
-				// Strip trailing /source from member paths — it's noise in summaries
-				cleaned := make([]string, len(top))
-				for i, m := range top {
-					cleaned[i] = strings.TrimSuffix(m, "/source")
-				}
-				sr.Communities = append(sr.Communities, communitySummary{
-					ID:         c.ID,
-					Size:       len(c.Members),
-					TopMembers: cleaned,
-				})
-			}
-			data, _ := json.MarshalIndent(sr, "", "  ")
-			return mcp.NewToolResultText(string(data)), nil
-		}
-
-		// Non-summary full output. Bead mache-9cd921: previously
-		// marshaled the full Communities slice, blowing past the MCP
-		// response budget (720K chars on 28K LOC). Cap to the top N
-		// largest communities + top N members within each. Truncation
-		// is in-place on the result; the marshaled JSON keeps its
-		// existing CommunityResult shape (PascalCase) so consumers
-		// that already parse this output don't break.
-		const maxCommunities = 25
-		const maxMembersPerCommunity = 20
-
-		sort.Slice(result.Communities, func(i, j int) bool {
-			return len(result.Communities[i].Members) > len(result.Communities[j].Members)
-		})
-		eliddedCommunities := 0
-		if len(result.Communities) > maxCommunities {
-			eliddedCommunities = len(result.Communities) - maxCommunities
-			result.Communities = result.Communities[:maxCommunities]
-		}
-		eliddedMembers := 0
-		for i := range result.Communities {
-			if len(result.Communities[i].Members) > maxMembersPerCommunity {
-				eliddedMembers += len(result.Communities[i].Members) - maxMembersPerCommunity
-				result.Communities[i].Members = result.Communities[i].Members[:maxMembersPerCommunity]
-			}
-		}
-
-		// Anonymous wrapper embeds *CommunityResult so existing
-		// consumers see the same field names; truncation metadata
-		// rides as additive optional fields.
-		out := struct {
-			*graph.CommunityResult
-			ElidedCommunities int    `json:",omitempty"`
-			ElidedMembers     int    `json:",omitempty"`
-			TruncationNote    string `json:",omitempty"`
-		}{
-			CommunityResult:   result,
-			ElidedCommunities: eliddedCommunities,
-			ElidedMembers:     eliddedMembers,
-		}
-		if eliddedCommunities > 0 || eliddedMembers > 0 {
-			out.TruncationNote = fmt.Sprintf(
-				"output truncated to %d communities (×%d members each) to fit MCP response budget; use summary=true for an even tighter view",
-				maxCommunities, maxMembersPerCommunity,
-			)
-		}
-		data, _ := json.MarshalIndent(out, "", "  ")
-		return mcp.NewToolResultText(string(data)), nil
-	}
-}
-
-// minFuzzyLen guards the partial-substring branch. Symbols shorter than
-// this trigger fuzzy matches against essentially every common token in
-// large codebases, which is exactly the noise mache-nmia complained
-// about. 4 chars is the smallest length where substring matching is
-// usually meaningful (e.g. "auth" → AuthZ, OAuth, authenticate).
-const minFuzzyLen = 4
-
-// maxFuzzySuggestions caps how many partial-match suggestions ride back
-// in a single response. The prior limit was 20 — still a wall of text in
-// monorepos. 8 keeps the suggestion list scannable.
-const maxFuzzySuggestions = 8
-
-func makeFindDefinitionHandler(g graph.Graph) server.ToolHandlerFunc {
-	return func(_ context.Context, request mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-		symbol := request.GetString("symbol", "")
-		if symbol == "" {
-			return mcp.NewToolResultError("symbol is required"), nil
-		}
-		fuzzy := request.GetBool("fuzzy", false)
-
-		// 1. Anchored exact match — case-sensitive. Use LookupDef
-		// when the backend supports it: O(1) map lookup, no
-		// O(N) snapshot copy of the entire defs map. Falls back
-		// to DefsMap for backends that haven't implemented the
-		// optional lookup method.
-		if dl, ok := g.(defsLookuper); ok {
-			if dirIDs := dl.LookupDef(symbol); len(dirIDs) > 0 {
-				if r := findDefinitionResultScoped(g, symbol, dirIDs); r != nil {
-					return r, nil
-				}
-				return findDefinitionResult(symbol, dirIDs), nil
-			}
-		}
-
-		// Backends that ONLY expose DefsMap (older/composite shapes)
-		// or that didn't hit on the LookupDef path fall through to
-		// the snapshot for the case-insensitive + fuzzy paths below.
-		dp := g.(defsMapProvider)
-		defs := dp.DefsMap()
-
-		// 1b. Anchored exact via the snapshot — only reached when
-		// the backend lacks a defsLookuper. Mirrors the old behavior.
-		if _, hasLookup := g.(defsLookuper); !hasLookup {
-			if dirIDs, ok := defs[symbol]; ok {
-				if r := findDefinitionResultScoped(g, symbol, dirIDs); r != nil {
-					return r, nil
-				}
-				return findDefinitionResult(symbol, dirIDs), nil
-			}
-		}
-
-		// 2. Anchored exact match — case-insensitive.
-		// Same token, just different case — still "anchored," not noise.
-		symbolLower := strings.ToLower(symbol)
-		for token, ids := range defs {
-			if strings.ToLower(token) == symbolLower {
-				if r := findDefinitionResultScoped(g, symbol, ids); r != nil {
-					return r, nil
-				}
-				return findDefinitionResult(symbol, ids), nil
-			}
-		}
-
-		// 3. Fuzzy substring fallback — opt-in only. The default behavior
-		// previously returned a wall of partial matches that drowned out
-		// the right answer in monorepos (bead mache-nmia).
-		if fuzzy && len(symbol) >= minFuzzyLen {
-			matches := collectFuzzyMatches(defs, symbolLower, maxFuzzySuggestions)
-			if len(matches) > 0 {
-				type suggestion struct {
-					Message     string   `json:"message"`
-					Suggestions []string `json:"suggestions"`
-				}
-				data, _ := json.MarshalIndent(suggestion{
-					Message:     fmt.Sprintf("no exact definition for %q, but found similar symbols (fuzzy=true)", symbol),
-					Suggestions: matches,
-				}, "", "  ")
-				return mcp.NewToolResultText(string(data)), nil
-			}
-		}
-
-		// 4. LSP fallback: try _lsp_defs from a ley-line pre-baked DB.
-		if qg, ok := g.(refsQuerier); ok {
-			lspDefs, err := queryLSPDefs(qg, symbol)
-			if err == nil && len(lspDefs) > 0 {
-				type lspResult struct {
-					Symbol      string           `json:"symbol"`
-					Source      string           `json:"source"`
-					Definitions []lspDefLocation `json:"definitions"`
-				}
-				data, _ := json.MarshalIndent(lspResult{
-					Symbol:      symbol,
-					Source:      "lsp",
-					Definitions: lspDefs,
-				}, "", "  ")
-				return mcp.NewToolResultText(string(data)), nil
-			}
-		}
-
-		// 5. Nothing matched. Hint at fuzzy if available.
-		hint := ""
-		if !fuzzy && len(symbol) >= minFuzzyLen {
-			hint = " — try fuzzy=true for substring matches"
-		}
-		if serveControl != "" {
-			return mcp.NewToolResultText(fmt.Sprintf("no definition found for %q — daemon may still be parsing, retry shortly%s", symbol, hint)), nil
-		}
-		return mcp.NewToolResultText(fmt.Sprintf("no definition found for %q%s", symbol, hint)), nil
-	}
-}
-
-// findDefinitionResult shapes the success response for find_definition.
-func findDefinitionResult(symbol string, dirIDs []string) *mcp.CallToolResult {
-	type defResult struct {
-		Symbol      string   `json:"symbol"`
-		Definitions []string `json:"definitions"`
-	}
-	data, _ := json.MarshalIndent(defResult{Symbol: symbol, Definitions: dirIDs}, "", "  ")
-	return mcp.NewToolResultText(string(data))
-}
-
-// findDefinitionResultScoped is the cross-repo variant: when the
-// active graph is a CompositeGraph and at least one definition
-// carries a mount prefix, emit the {symbol, definitions: [{path,
-// mount}]} shape so agents see which mount each def came from.
-// Returns nil if annotation adds nothing — caller falls through to
-// findDefinitionResult.
-func findDefinitionResultScoped(g graph.Graph, symbol string, dirIDs []string) *mcp.CallToolResult {
-	scoped := annotateMounts(g, dirIDs)
-	if scoped == nil {
-		return nil
-	}
-	type defResult struct {
-		Symbol      string       `json:"symbol"`
-		Definitions []scopedItem `json:"definitions"`
-	}
-	data, _ := json.MarshalIndent(defResult{Symbol: symbol, Definitions: scoped}, "", "  ")
-	return mcp.NewToolResultText(string(data))
-}
-
-// collectFuzzyMatches walks the defs map looking for tokens whose
-// case-insensitive form contains symbolLower (or vice versa) and returns
-// up to `limit` "token → dir_id" pairs. Caller has already ensured
-// symbolLower is at least minFuzzyLen.
-func collectFuzzyMatches(defs map[string][]string, symbolLower string, limit int) []string {
-	var matches []string
-	for token, ids := range defs {
-		if len(matches) >= limit {
-			break
-		}
-		tokenLower := strings.ToLower(token)
-		if !strings.Contains(tokenLower, symbolLower) && !strings.Contains(symbolLower, tokenLower) {
-			continue
-		}
-		for _, id := range ids {
-			matches = append(matches, token+" → "+id)
-			if len(matches) >= limit {
-				break
-			}
-		}
-	}
-	return matches
-}
-
-func makeGetOverviewHandler(g graph.Graph) server.ToolHandlerFunc {
-	return func(ctx context.Context, request mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-		type dirInfo struct {
-			Name     string `json:"name"`
-			Path     string `json:"path"`
-			Children int    `json:"children"`
-		}
-		type overview struct {
-			TopLevel   []dirInfo         `json:"top_level"`
-			TotalDirs  int               `json:"total_dirs"`
-			TotalFiles int               `json:"total_files"`
-			RefTokens  int               `json:"ref_tokens,omitempty"`
-			DefTokens  int               `json:"def_tokens,omitempty"`
-			Usage      map[string]string `json:"_usage,omitempty"`
-		}
-
-		ov := overview{}
-
-		// Top-level structure
-		children, err := g.ListChildren("")
-		if err != nil {
-			return mcp.NewToolResultError(fmt.Sprintf("list root: %v", err)), nil
-		}
-
-		for _, childID := range children {
-			node, err := g.GetNode(childID)
-			if err != nil {
-				continue
-			}
-			if node.Mode.IsDir() {
-				ov.TotalDirs++
-				subChildren, _ := g.ListChildren(childID)
-				ov.TopLevel = append(ov.TopLevel, dirInfo{
-					Name:     filepath.Base(childID),
-					Path:     childID,
-					Children: len(subChildren),
-				})
-			} else {
-				ov.TotalFiles++
-			}
-		}
-
-		// Count refs/defs if available
-		if rp, ok := g.(refsMapProvider); ok {
-			ov.RefTokens = len(rp.RefsMap())
-		}
-		if dp, ok := g.(defsMapProvider); ok {
-			ov.DefTokens = len(dp.DefsMap())
-		}
-
-		// Embed tool routing hints when a code graph is indexed (has cross-references).
-		// This teaches the LLM which tool to use for each task without bloating the
-		// system prompt — guidance is only in context after get_overview is called.
-		if ov.RefTokens > 0 {
-			ov.Usage = map[string]string{
-				"find_callers":    "who calls a symbol — use instead of grep for 'who uses X?'",
-				"find_definition": "where a symbol is declared — use for 'where is X defined?'",
-				"find_callees":    "what a function invokes — note: generic names (String, New, Error) may have false positives",
-				"search":          "find symbols by name pattern, e.g. '%auth%' or 'Parse%' — use instead of grep -r",
-				"list_directory":  "browse the tree structure — use instead of ls/find",
-				"get_communities": "find clusters of related code (use summary=true for large repos; requires dense cross-references)",
-				"get_impact":      "blast radius of changing a symbol — traces callers/callees to a configurable depth",
-			}
-		}
-
-		// Walk one level deeper to count total dirs/files
-		for _, childID := range children {
-			node, _ := g.GetNode(childID)
-			if node != nil && node.Mode.IsDir() {
-				subChildren, _ := g.ListChildren(childID)
-				for _, subID := range subChildren {
-					subNode, _ := g.GetNode(subID)
-					if subNode != nil {
-						if subNode.Mode.IsDir() {
-							ov.TotalDirs++
-						} else {
-							ov.TotalFiles++
-						}
-					}
-				}
-			}
-		}
-
-		data, _ := json.MarshalIndent(ov, "", "  ")
-		return mcp.NewToolResultText(string(data)), nil
-	}
-}
-
-// makeGetSheafStatusHandler returns the handler for the get_sheaf_status
-// MCP tool. The tool surfaces the daemon's sheaf cache state to agents
-// so they can decide whether cached results are still fresh.
-//
-// Design contract: the handler MUST NOT surface daemon unavailability
-// as an MCP error. Agents polling this tool would otherwise see
-// transport failures whenever the daemon is down or hasn't been
-// dialed yet. Instead, return a structured {available: false, reason:
-// "..."} response. This matches the documented graceful-degradation
-// pattern in the wider sheaf wiring (cmd/serve.go).
-//
-// DiscoverSocket (not DiscoverOrStart) is the right primitive here:
-// a status check should never trigger a daemon auto-spawn (which can
-// download the binary, take seconds, etc.).
-func makeGetSheafStatusHandler() server.ToolHandlerFunc {
-	return func(ctx context.Context, request mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-		unavailable := func(reason string) (*mcp.CallToolResult, error) {
-			data, _ := json.Marshal(map[string]any{
-				"available": false,
-				"reason":    reason,
-			})
-			return mcp.NewToolResultText(string(data)), nil
-		}
-
-		sockPath, err := leyline.DiscoverSocket()
-		if err != nil {
-			return unavailable("no ley-line daemon socket — set LEYLINE_SOCKET or start `leyline daemon`")
-		}
-		sock, err := leyline.DialSocket(sockPath)
-		if err != nil {
-			return unavailable(fmt.Sprintf("dial %s: %v", sockPath, err))
-		}
-		defer func() { _ = sock.Close() }()
-
-		sc := leyline.NewSheafClient(sock)
-		s, err := sc.Status()
-		if err != nil {
-			return unavailable(fmt.Sprintf("sheaf_status: %v", err))
-		}
-
-		data, _ := json.Marshal(map[string]any{
-			"available":  true,
-			"generation": s.Generation,
-			"valid":      s.Valid,
-			"total":      s.Total,
-			"defect":     s.Defect,
-		})
-		return mcp.NewToolResultText(string(data)), nil
-	}
 }
